api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
python-multipart==0.0.20
email-validator==2.3.0
eval-type-backport==0.3.1
psycopg2-binary==2.9.11
